change datatype of weight

+++ src/main/java/com/team/controller/InsertPetInfoController.java
@@ -26,10 +26,9 @@ public class InsertPetInfoController {
     public ResponseEntity<?> insertPetInfo(@RequestBody Map<String, String> data) {
 
         String petName = data.get("petName");
-        float weight = Float.parseFloat(data.get("weight"));
+        double weight = Double.parseDouble(data.get("weight"));
         int age = Integer.parseInt(data.get("age"));
         Integer customerID = Integer.parseInt(data.get("customerID"));
-
         Customers customer = petDetailService.findCustomerById(customerID);
 
         // Check if pet name already exists in the database for the customer
